AutoBlog_Agent/requirements.txt
google-generativeai
loguru
requests
schedule
python-dotenv
beautifulsoup4
lxml
flask
flask-cors
groq
